Write the first integration test
... to validate the giornata-today interactive function.

The integration test suite is deterministic - it uses faketime to fake the time
- and self-destructive - it creates an ephemeral giornata-directory and attempts
to do thing in-memory (though this is only guaranteed if the underlying host's
/tmp is a tmpfs).

Next up: Instead of faketime though, we should use libfaketime to manipulate the
library from giornata-tests.el directly. This means we can do things like faking
the time any amount of times to populate a giornata-directory of more than one
entry.