Change giornata-front-matter variable type to a simple string
The previous type, an alist, elludes to the wrong assumption that
Giornata can utilize multiple front matters at once. It can't. What the
user can do is make this variable directory-local and customize it to
their liking.

I've added extensive documentation to help replace the gap left by this
removal.